requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bitarray==2.7.5
boto3==1.26.100
botocore==1.29.158
certifi==2023.5.7
charset-normalizer==3.1.0
cloudscraper==1.2.71
colorama==0.4.6
cytoolz==0.12.1
Django==4.1.7
django-ckeditor==6.5.1
django-cleanup==7.0.0
django-cors-headers==3.14.0
django-csp==3.7
django-filter==22.1
django-js-asset==2.0.0
django-nested-inline==0.4.6
django-otp==1.2.2
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
eth-abi==4.1.0
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.4.0
eth-utils==2.1.1
exceptiongroup==1.1.3
frozenlist==1.3.3
future==0.18.3
gunicorn==20.1.0
h11==0.14.0
hexbytes==0.3.1
httpcore==0.17.3
httpx==0.24.1
idna==3.4
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.17.3
lru-dict==1.2.0
multidict==6.0.4
netaddr==0.8.0
packaging==24.0
parsimonious==0.9.0
Pillow==9.4.0
protobuf==4.23.3
psycopg2-binary==2.9.5
pycryptodome==3.18.0
PyJWT==2.6.0
pyparsing==3.1.1
pypng==0.20220715.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-http-client==3.3.7
python-telegram-bot==20.4
pytz==2022.7.1
pywin32==306
PyYAML==6.0.1
regex==2023.6.3
requests==2.28.2
requests-toolbelt==1.0.0
rlp==3.0.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
toolz==0.12.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.16
varint==1.0.2
web3==6.0.0
websockets==11.0.3
xlwt==1.3.0
yarl==1.9.2
